+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.3 (2019-01-13)
+
+- Added AMP support
+- Updated support for custom HTML blocks (created with the Code Editor) by parsing CSS to JSON (just like a userblock.js and project.mobirise stores it)
+- Updated opening the Gallery. The version of the extension and is sent, along with what template the user is currently using (Bootstrap or AMP). This can be used by the Gallery
+- Added animated 'waiting' icon after activating a block. This is more clear to the user
+
 ## 0.2 (2019-01-06)
 
 - Changed template JSON format. The extension now only accepts 'user block' format. This makes it even easier to exchange blocks.